requirements.txt
python-dotenv
requests
google-adk
google-adk[a2a]
litellm
uvicorn